feat(persona): add actions and disable powers for Mare unit
This commit introduces a new persona for the 'Mare' unit, including a set
of actions such as stomping, curtseying, and playing various sound
effects. Additionally, it disables several powers for the unit, including
mind, inventory, flight, touch-self, far-cam, reach, amplifier, and
location. It also enables the 'bolts' policy and disables the 'release'
policy. Finally, it sets the unit's name